Finished moderation commands
- Added icons to all messages.
Success - ✔️
Error - ❌
Info - 🔵
Warning/Processing - ⚠️

- Added set, current and clear for modchat to mod.js
- Added set, current and clear for trustrole to mod.js
- Added set, current and clear for max difference to mod.js
- Added set, and current for moderation level to mod.js

- Changed max difference in ModControl table from bigint to float.
- Added Role_Name to Roles table.

+++ SQL_Files/MySQL/ModControl.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS `ModControl`
 (
  `Mod_Id`                int NOT NULL AUTO_INCREMENT ,
  `ModLevel_Id`           int NOT NULL ,
- `Stats_Update_Max_Diff` bigint NULL ,
+ `Stats_Update_Max_Diff` float NULL ,
  `Stats_Trusted_Role`    int NULL ,
  `Mod_Approval_Chat`     int NULL ,
  `Moderator_Role`        int NULL ,
